requirements.txt
protobuf~=5.29.4
yaml~=0.2.5
pyyaml~=6.0.2
numpy~=2.1.3
tensorflow~=2.19.0
pillow~=11.1.0
metaflow~=2.15.7
pylint~=3.3.7